+++ join.go
@@ -88,6 +88,8 @@ func (n *JoinNode) Point(src int, p edge.PointMessage) error {
 }
 
 func (n *JoinNode) Barrier(src int, b edge.BarrierMessage) error {
+	g := n.getOrCreateGroup(b.GroupID())
+	g.Barrier(src, b.Time())
 	return edge.Forward(n.outs, b)
 }
 
@@ -337,13 +339,26 @@ func (g *joinGroup) Collect(src int, p timeMessage) error {
 	// Update head
 	g.head[src] = t
 
-	onlyReadySets := false
-	for _, t := range g.head {
-		if !t.After(g.oldestTime) {
-			onlyReadySets = true
-			break
-		}
+	onlyReadySets := g.checkOnlyReadSets()
+	err := g.emit(onlyReadySets)
+	if err != nil {
+		return err
 	}
+	return nil
+}
+
+// Barrier signals a src will not produce points older than time.
+// Emit the oldest set if we have collected enough data.
+func (g *joinGroup) Barrier(src int, t time.Time) error {
+	t = t.Round(g.n.j.Tolerance)
+	if t.Before(g.oldestTime) || g.oldestTime.IsZero() {
+		g.oldestTime = t
+	}
+
+	// Update head
+	g.head[src] = t
+
+	onlyReadySets := g.checkOnlyReadSets()
 	err := g.emit(onlyReadySets)
 	if err != nil {
 		return err
@@ -367,6 +382,9 @@ func (g *joinGroup) newJoinset(t time.Time) *joinset {
 
 // emit a set and update the oldestTime.
 func (g *joinGroup) emit(onlyReadySets bool) error {
+	if len(g.sets) == 0 {
+		return nil
+	}
 	sets := g.sets[g.oldestTime]
 	i := 0
 	for ; i < len(sets); i++ {
@@ -391,7 +409,29 @@ func (g *joinGroup) emit(onlyReadySets bool) error {
 			g.oldestTime = t
 		}
 	}
+	// Check if there are more non ready sets we can emit.
+	// This occurs when one of the parents missed a section of data
+	// while the other parents continued on.
+	// We need to emit all the buffered sets as soon as all the parent heads have passed the oldesttime.
+	if !onlyReadySets {
+		onlyReadySets = g.checkOnlyReadSets()
+		return g.emit(onlyReadySets)
+	}
 	return nil
+}
+
+// checkOnlyReadSets reports if all heads are past the oldesttime,
+// indicated whether its ok to emit non ready sets.
+func (g *joinGroup) checkOnlyReadSets() bool {
+	onlyReadySets := false
+	// Check if heads are past oldest time
+	for _, t := range g.head {
+		if !t.After(g.oldestTime) {
+			onlyReadySets = true
+			break
+		}
+	}
+	return onlyReadySets
 }
 
 // emit sets until we have none left.