+++ server/server_test.go
@@ -6671,12 +6671,12 @@ func TestServer_UpdateConfig(t *testing.T) {
 				Elements: []client.ConfigElement{{
 					Link: client.Link{Relation: client.Self, Href: "/kapacitor/v1/config/alerta/"},
 					Options: map[string]interface{}{
-						"enabled":      false,
-						"environment":  "",
-						"origin":       "",
-						"token":        false,
-						"token-prefix": "",
-						"url":          "http://alerta.example.com",
+						"enabled":              false,
+						"environment":          "",
+						"origin":               "",
+						"token":                false,
+						"token-prefix":         "",
+						"url":                  "http://alerta.example.com",
 						"insecure-skip-verify": false,
 						"timeout":              "0s",
 					},
@@ -6688,12 +6688,12 @@ func TestServer_UpdateConfig(t *testing.T) {
 			expDefaultElement: client.ConfigElement{
 				Link: client.Link{Relation: client.Self, Href: "/kapacitor/v1/config/alerta/"},
 				Options: map[string]interface{}{
-					"enabled":      false,
-					"environment":  "",
-					"origin":       "",
-					"token":        false,
-					"token-prefix": "",
-					"url":          "http://alerta.example.com",
+					"enabled":              false,
+					"environment":          "",
+					"origin":               "",
+					"token":                false,
+					"token-prefix":         "",
+					"url":                  "http://alerta.example.com",
 					"insecure-skip-verify": false,
 					"timeout":              "0s",
 				},
@@ -6715,12 +6715,12 @@ func TestServer_UpdateConfig(t *testing.T) {
 						Elements: []client.ConfigElement{{
 							Link: client.Link{Relation: client.Self, Href: "/kapacitor/v1/config/alerta/"},
 							Options: map[string]interface{}{
-								"enabled":      false,
-								"environment":  "",
-								"origin":       "kapacitor",
-								"token":        true,
-								"token-prefix": "",
-								"url":          "http://alerta.example.com",
+								"enabled":              false,
+								"environment":          "",
+								"origin":               "kapacitor",
+								"token":                true,
+								"token-prefix":         "",
+								"url":                  "http://alerta.example.com",
 								"insecure-skip-verify": false,
 								"timeout":              "3h0m0s",
 							},
@@ -6732,12 +6732,12 @@ func TestServer_UpdateConfig(t *testing.T) {
 					expElement: client.ConfigElement{
 						Link: client.Link{Relation: client.Self, Href: "/kapacitor/v1/config/alerta/"},
 						Options: map[string]interface{}{
-							"enabled":      false,
-							"environment":  "",
-							"origin":       "kapacitor",
-							"token":        true,
-							"token-prefix": "",
-							"url":          "http://alerta.example.com",
+							"enabled":              false,
+							"environment":          "",
+							"origin":               "kapacitor",
+							"token":                true,
+							"token-prefix":         "",
+							"url":                  "http://alerta.example.com",
 							"insecure-skip-verify": false,
 							"timeout":              "3h0m0s",
 						},
@@ -8795,6 +8795,16 @@ func TestServer_ListServiceTests(t *testing.T) {
 						"Category":    "",
 					},
 					"timestamp": "2014-11-12T11:45:26.371Z",
+					"links": []interface{}{
+						map[string]interface{}{
+							"href": "https://example.com/a",
+							"text": "a",
+						},
+						map[string]interface{}{
+							"href": "https://example.com/b",
+							"text": "b",
+						},
+					},
 				},
 			},
 			{
@@ -8826,6 +8836,7 @@ func TestServer_ListServiceTests(t *testing.T) {
 					"output":   "testOutput",
 					"source":   "Kapacitor",
 					"handlers": []interface{}{},
+					"metadata": map[string]interface{}{},
 					"level":    "CRITICAL",
 				},
 			},
@@ -8965,6 +8976,7 @@ func TestServer_ListServiceTests_WithPattern(t *testing.T) {
 					"output":   "testOutput",
 					"source":   "Kapacitor",
 					"handlers": []interface{}{},
+					"metadata": map[string]interface{}{},
 					"level":    "CRITICAL",
 				},
 			},
@@ -9856,6 +9868,16 @@ func TestServer_AlertHandlers(t *testing.T) {
 				Kind: "pagerduty2",
 				Options: map[string]interface{}{
 					"routing-key": "rkey",
+					"links": []interface{}{
+						map[string]string{
+							"href": "http://example.com",
+							"text": "t1",
+						},
+						map[string]string{
+							"href": "http://example.com/{{.TaskName}}",
+							"text": "t2",
+						},
+					},
 				},
 			},
 			setup: func(c *server.Config, ha *client.TopicHandler) (context.Context, error) {
@@ -9899,6 +9921,10 @@ func TestServer_AlertHandlers(t *testing.T) {
 							Timestamp: "1970-01-01T00:00:00.000000000Z",
 						},
 						RoutingKey: "rkey",
+						Links: []pagerduty2test.Link{
+							{Href: "http://example.com", Text: "t1"},
+							{Href: "http://example.com/testAlertHandlers", Text: "t2"},
+						},
 					},
 				}}
 
@@ -10000,6 +10026,10 @@ func TestServer_AlertHandlers(t *testing.T) {
 				Kind: "sensu",
 				Options: map[string]interface{}{
 					"source": "Kapacitor",
+					"metadata": map[string]interface{}{
+						"k1": "v1",
+						"k2": 5,
+					},
 				},
 			},
 			setup: func(c *server.Config, ha *client.TopicHandler) (context.Context, error) {
@@ -10022,6 +10052,10 @@ func TestServer_AlertHandlers(t *testing.T) {
 					Output: "message",
 					Name:   "id",
 					Status: 2,
+					Metadata: map[string]interface{}{
+						"k1": "v1",
+						"k2": float64(5),
+					},
 				}}
 				got := ts.Requests()
 				if !reflect.DeepEqual(exp, got) {
@@ -10105,9 +10139,9 @@ func TestServer_AlertHandlers(t *testing.T) {
 						"Mime-Version":              []string{"1.0"},
 						"Content-Type":              []string{"text/html; charset=UTF-8"},
 						"Content-Transfer-Encoding": []string{"quoted-printable"},
-						"To":      []string{"oncall@example.com, backup@example.com"},
-						"From":    []string{"test@example.com"},
-						"Subject": []string{"message"},
+						"To":                        []string{"oncall@example.com, backup@example.com"},
+						"From":                      []string{"test@example.com"},
+						"Subject":                   []string{"message"},
 					},
 					Body: "details\n",
 				}}