+++ services/httppost/service.go
@@ -2,6 +2,7 @@ package httppost
 
 import (
 	"bytes"
+	"crypto/tls"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -14,7 +15,9 @@ import (
 	"time"
 
 	"context"
+
 	"github.com/influxdata/kapacitor/alert"
+	khttp "github.com/influxdata/kapacitor/http"
 	"github.com/influxdata/kapacitor/keyvalue"
 	"github.com/pkg/errors"
 )
@@ -233,11 +236,12 @@ func (s *Service) Test(options interface{}) error {
 }
 
 type HandlerConfig struct {
-	URL             string            `mapstructure:"url"`
-	Endpoint        string            `mapstructure:"endpoint"`
-	Headers         map[string]string `mapstructure:"headers"`
-	CaptureResponse bool              `mapstructure:"capture-response"`
-	Timeout         time.Duration     `mapstructure:"timeout"`
+	URL                 string            `mapstructure:"url"`
+	Endpoint            string            `mapstructure:"endpoint"`
+	Headers             map[string]string `mapstructure:"headers"`
+	CaptureResponse     bool              `mapstructure:"capture-response"`
+	Timeout             time.Duration     `mapstructure:"timeout"`
+	SkipSSLVerification bool              `mapstructure:"skip-ssl-verification"`
 }
 
 type handler struct {
@@ -252,6 +256,8 @@ type handler struct {
 
 	timeout time.Duration
 
+	skipSSLVerification bool
+
 	hc *http.Client
 }
 
@@ -261,12 +267,13 @@ func (s *Service) Handler(c HandlerConfig, ctx ...keyvalue.T) alert.Handler {
 		e = NewEndpoint(c.URL, nil, BasicAuth{}, nil, nil)
 	}
 	return &handler{
-		s:               s,
-		endpoint:        e,
-		diag:            s.diag.WithContext(ctx...),
-		headers:         c.Headers,
-		captureResponse: c.CaptureResponse,
-		timeout:         c.Timeout,
+		s:                   s,
+		endpoint:            e,
+		diag:                s.diag.WithContext(ctx...),
+		headers:             c.Headers,
+		captureResponse:     c.CaptureResponse,
+		timeout:             c.Timeout,
+		skipSSLVerification: c.SkipSSLVerification,
 	}
 }
 
@@ -323,8 +330,20 @@ func (h *handler) Handle(event alert.Event) {
 		req = req.WithContext(ctx)
 	}
 
+	// Setup HTTP client
+	var tlsConfig *tls.Config
+	if h.skipSSLVerification {
+		tlsConfig = &tls.Config{
+			InsecureSkipVerify: true,
+		}
+	}
+
+	httpClient := &http.Client{
+		Transport: khttp.NewDefaultTransportWithTLS(tlsConfig),
+	}
+
 	// Execute the request
-	resp, err := http.DefaultClient.Do(req)
+	resp, err := httpClient.Do(req)
 	if err != nil {
 		h.diag.Error("failed to POST alert data", err)
 		return